Update CLI LaaS for integration
