+++ bin/log/index.js
@@ -19,7 +19,6 @@ const credential_type = ['password'];
 
 category.addChild(require('./create')(resource));
 category.addChild(require('./stream')(resource));
-category.addChild(require('./read')(resource));
 category.addChild(require('../generic/credential')(resource, credential_type));
 
 module.exports = category;
