--- bin/log/read/examples.md
@@ -1,3 +1,0 @@
-```bash
-{{command_name}} --log my-server-log
-```
